Moved data structure types into new package com.valgolang.frontend.datastructures
