+++ src/main/kotlin/com/valgolang/runtime/datastructures/stack/StackExecutor.kt
@@ -1,6 +1,10 @@
 package com.valgolang.runtime.datastructures.stack
 
-import com.valgolang.frontend.*
+import com.valgolang.frontend.AssignLHS
+import com.valgolang.frontend.ConstructorNode
+import com.valgolang.frontend.DataStructureInitialiserNode
+import com.valgolang.frontend.MethodCallNode
+import com.valgolang.frontend.datastructures.stack.StackType
 import com.valgolang.linearrepresentation.*
 import com.valgolang.linearrepresentation.datastructures.stack.InitManimStack
 import com.valgolang.linearrepresentation.datastructures.stack.StackPopObject
