+++ src/test/kotlin/com/valgolang/ASTConstructionTests.kt
@@ -1,6 +1,8 @@
 package com.valgolang
 
 import com.valgolang.frontend.*
+import com.valgolang.frontend.datastructures.binarytree.NodeType
+import com.valgolang.frontend.datastructures.stack.StackType
 import junit.framework.TestCase.assertEquals
 import org.junit.jupiter.api.Test
